Added user preference and location commands, restructered files
